requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
pillow

# tests/linter
#black
#coverage
#flake8
#pytest
#yapf

# API
#fastapi
#uvicorn
streamlit

# utilities
#six>=1.14
#joblib
#memoized-property
#termcolor
python-dotenv
